Fixed RSS feed
The RSS feed for this site did not contain the full content
of the articles in this project on previous commits due to
an ongoing issue with the {distill} package. To address this,
the Rmarkdown files in this projet were rendered again after
installing the development version of distill using
'devtools::install_github("rstudio/distill")'.

While rendering the articles again using this development
version of {distill}, errors arose with 'ggplot2::ggsave()'
commands and with the relative paths to these saved files.
The ggsave() commands were removed as they were no longer
needed and the file paths were fixed. The ggsave() commands
were redundant as the plots are already written to files
during rendering of distill articles using
'rmarkdown::render()'.

This commit was delayed due to issues installing the {sf}
package used in the US Post Offices article, despite
having met the system requirements outlined on CRAN.
The package was successfully installed once the path to
the system libraries was specified during the installation
using an additional argument. For example:
'install.packages("sf",
                  configure.args = "--with-proj-lib=/usr:/lib")'